add optimization for double/float/boolean in PixelsPageSource

+++ pixels-core/src/main/java/cn/edu/ruc/iir/pixels/core/utils/BitUtils.java
@@ -1,7 +1,6 @@
 package cn.edu.ruc.iir.pixels.core.utils;
 
 import java.io.ByteArrayOutputStream;
-import java.nio.ByteBuffer;
 
 /**
  * pixels
@@ -23,7 +22,7 @@ public class BitUtils
 
         for (int i = 0; i < length; i++)
         {
-            long v = values[i] ? 1 : 0;
+            byte v = values[i] ? (byte) 1 : (byte) 0;
             bitsLeft -= bitsToWrite;
             current |= v << bitsLeft;
             if (bitsLeft == 0)
@@ -42,7 +41,35 @@ public class BitUtils
         return bitWiseOutput.toByteArray();
     }
 
-    public static byte[] bitWiseCompact(long[] values, int length)
+//    public static byte[] bitWiseCompact(long[] values, int length)
+//    {
+//        ByteArrayOutputStream bitWiseOutput = new ByteArrayOutputStream();
+//        int bitsToWrite = 1;
+//        int bitsLeft = 8;
+//        byte current = 0;
+//
+//        for (int i = 0; i < length; i++)
+//        {
+//            long v = values[i];
+//            bitsLeft -= bitsToWrite;
+//            current |= v << bitsLeft;
+//            if (bitsLeft == 0)
+//            {
+//                bitWiseOutput.write(current);
+//                current = 0;
+//                bitsLeft = 8;
+//            }
+//        }
+//
+//        if (bitsLeft != 8)
+//        {
+//            bitWiseOutput.write(current);
+//        }
+//
+//        return bitWiseOutput.toByteArray();
+//    }
+
+    public static byte[] bitWiseCompact(byte[] values, int length)
     {
         ByteArrayOutputStream bitWiseOutput = new ByteArrayOutputStream();
         int bitsToWrite = 1;
@@ -51,7 +78,7 @@ public class BitUtils
 
         for (int i = 0; i < length; i++)
         {
-            long v = values[i];
+            byte v = values[i];
             bitsLeft -= bitsToWrite;
             current |= v << bitsLeft;
             if (bitsLeft == 0)
